requirements.txt
easyocr
numpy
pandas
Pillow
pydantic
Requests
SQLAlchemy
streamlit
streamlit_paste_button
streamlit-paste-button
sqlalchemy
